feat(backend): Migrar sistema de stock de Integer a Double para soporte decimal  MODELOS DE DOMINIO: - ArticuloInsumo: stockActual, stockMinimo (Integer → Double) - ArticuloManufacturadoDetalle: cantidad (Integer → Double) - DetallePedido: cantidad (Integer → Double) - PromocionDetalle: cantidadRequerida (Integer → Double) - CompraDetalle: cantidad (Integer → Double)  DTOs ACTUALIZADOS: - ProduccionAnalisisDTO: cantidadNecesaria, cantidadDisponible (Integer → Double) - CompraDetalleDTO: cantidad (Integer → Double) - ArticuloInsumoDTO: stockActual, stockMinimo (Integer → Double) - IngredienteDTO: cantidad (Integer → Double) - ArticuloManufacturadoDetalleDTO: cantidad (Integer → Double) - DetallePedidoDTO: cantidad (Integer → Double)  SERVICIOS MODIFICADOS: - ArticuloInsumoService:    * Validaciones de stock con Double   * Cálculos de disponibilidad con precisión decimal - StockService:   * Operaciones de suma/resta con valores decimales   * Validaciones de stock mínimo con Double - ArticuloValidacionService:   * Verificación de cantidades disponibles con decimales - HistorialService:   * Registro de movimientos con cantidades decimales - ProduccionService:   * Cálculo de ingredientes necesarios con precisión decimal  CONTROLADORES: - ArticuloController: Endpoints actualizados para recibir/devolver Double - CompraController: Manejo de cantidades decimales en compras - PedidoController: Procesamiento de cantidades decimales en pedidos  CONFIGURACIÓN DE DATOS: - DataInitializer:    * Stocks de insumos para elaborar: valores con .00 (ej: 10000.00)   * Stocks de productos terminados: valores exactos (ej: 300.00)   * Cantidades en recetas: precisión decimal (ej: 200.00g harina)   * Promociones corregidas con persistencia adecuada  VALIDACIONES MEJORADAS: - Verificación de stock con precisión decimal - Cálculos de costos con mayor exactitud - Validaciones de cantidades mínimas para compras - Control de stock crítico con decimales  BASE DE DATOS: - Migración de columnas INTEGER a DOUBLE/DECIMAL - Mantenimiento de precisión en operaciones aritméticas - Índices actualizados para nuevos tipos de datos  IMPACTO TÉCNICO: + Precisión en control de inventario (2.5kg vs 2kg) + Cálculos exactos en recetas y costos + Soporte para unidades fraccionarias (litros, kilos) + Eliminación de errores de redondeo en operaciones + Base sólida para reportes financieros precisos  COMPATIBILIDAD: - Migración backward-compatible - Datos existentes preservados con conversión automática - APIs mantienen contratos con nuevos tipos

+++ src/main/java/utn/saborcito/El_saborcito_back/controllers/ArticuloController.java
@@ -80,12 +80,11 @@ public class ArticuloController {
     @PostMapping("/analizar-produccion")
     public ProduccionAnalisisDTO analizarProduccion(@RequestBody ProduccionAnalisisRequestDTO requestDTO) {
         // Convertir la lista de objetos ArticuloCantidad a un mapa de ID -> cantidad
-        Map<Long, Integer> articulosMap = new HashMap<>();
+        Map<Long, Double> articulosMap = new HashMap<>();    // ✅ Cambio Integer a Double
 
         for (ProduccionAnalisisRequestDTO.ArticuloCantidad articulo : requestDTO.getArticulos()) {
-            // Si ya existe el ID en el mapa, sumar las cantidades (para artículos
-            // repetidos)
-            articulosMap.merge(articulo.getArticuloId(), articulo.getCantidad(), Integer::sum);
+            // Si ya existe el ID en el mapa, sumar las cantidades (para artículos repetidos)
+            articulosMap.merge(articulo.getArticuloId(), articulo.getCantidad(), Double::sum);    // ✅ Cambio Integer::sum a Double::sum
         }
 
         // Llamar al servicio unificado que hace el análisis real
